Update director.py
Line 103 seems to work, but line 102 doesn't. Tried the other tmx files in the repository instead of "adventure_level.tms" and they also don't seem to work. 
If we can't figure out a solution on how to get past the menu screen today, let's just use line 103 and figure it out next week. We'll still be fulfilling all the Beta requirements for this week.